+++ CHANGELOG.md
@@ -79,6 +79,8 @@
     - `DOCKER_BASE_IMAGE_REPO` - just the repo name
     - `DOCKER_BASE_IMAGE_TAG` - just the tag
     - `DOCKER_BASE_IMAGE_DIGEST` - just the digest
+  - `_IMAGE_DIGEST` - docker registry v2 image manifest digest
+  - `_IMAGE_LIST_DIGEST` - docker registry v2 image list manifest digest
   - `_SIGNATURES` - all docker registry cosign signatures
   - All `uname()` fields have dedicated fields:
     - `HOST_SYSNAME_WHEN_CHALKED`